+++ tools/utils/buildmessage.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var files = require('../fs/files');
-var parseStack = require('./parse-stack.js');
+var parseStack = require('./parse-stack');
 var fiberHelpers = require('./fiber-helpers.js');
 var Progress = require('../console/progress').Progress;
