+++ tools/utils/parse-stack.ts
